Volume update: Filter fixtures and v1
