+++ lib/fog/openstack/volume/v1/models/volume.rb
@@ -17,7 +17,8 @@ module Fog
             data = if id.nil?
                      service.create_volume(display_name, display_description, size, attributes)
                    else
-                     service.update_volume(id, attributes.reject { |k, _v| k == :id })
+                     attrib = attributes.select { |key| %i(display_name display_description metadata).include?(key) }
+                     service.update_volume(id, attrib)
                    end
             merge_attributes(data.body['volume'])
             true
